fix: FakeSparseEmbedding is now defined in an unique file conftest.py and imported in the 5 other test files. Centralised DOCS as SAMPLE_DOCS in conftest.py too. Corrected test_rag.py if cuda is not available.
