+++ tests/test_rag_api.py
@@ -1,4 +1,3 @@
-from typing import Dict, List
 from unittest.mock import patch
 
 import pytest
@@ -7,34 +6,19 @@ from fastapi.testclient import TestClient
 from langchain_core.language_models.fake_chat_models import FakeListChatModel
 from langchain_core.prompts import ChatPromptTemplate
 from langchain_community.embeddings import FakeEmbeddings
-from langchain_milvus.utils.sparse import BaseSparseEmbedding
 from pydantic import BaseModel, Field
 from pymilvus import MilvusClient
 
+from conftest import SAMPLE_DOCS, FakeSparseEmbedding
 from mmore.index.indexer import Indexer
 from mmore.rag.model import DenseModelConfig, SparseModelConfig
 from mmore.rag.pipeline import DEFAULT_PROMPT, RAGPipeline
 from mmore.rag.retriever import Retriever
-from mmore.type import MultimodalSample
-
-
-class _FakeSparseEmbedding(BaseSparseEmbedding):
-    def embed_query(self, query: str) -> Dict[int, float]:
-        return {0: 1.0, 1: float(len(query))}
-
-    def embed_documents(self, texts: List[str]) -> List[Dict[int, float]]:
-        return [{0: 1.0, i + 1: float(len(t))} for i, t in enumerate(texts)]
 
 
 class RAGInput(BaseModel):
     input: str = Field(..., description="The user query or question.")
     collection_name: str = Field(..., description="The Milvus collection name to search.")
-
-
-_DOCS = [
-    MultimodalSample(id="doc-1", document_id="doc-1", text="Paris is the capital of France.", modalities=[], metadata={}),
-    MultimodalSample(id="doc-2", document_id="doc-2", text="The Eiffel Tower stands 330 metres tall.", modalities=[], metadata={}),
-]
 
 _COLLECTION = "test_col"
 
@@ -42,14 +26,14 @@ _COLLECTION = "test_col"
 @pytest.fixture(scope="module")
 def populated_db(tmp_path_factory):
     db_path = str(tmp_path_factory.mktemp("api_db") / "test.db")
-    with patch("mmore.index.indexer.SparseModel.from_config", return_value=_FakeSparseEmbedding()):
+    with patch("mmore.index.indexer.SparseModel.from_config", return_value=FakeSparseEmbedding()):
         client = MilvusClient(db_path, enable_sparse=True)
         indexer = Indexer(
             dense_model_config=DenseModelConfig(model_name="debug"),
             sparse_model_config=SparseModelConfig(model_name="naver/splade-cocondenser-selfdistil"),
             client=client,
         )
-        indexer.index_documents(_DOCS, collection_name=_COLLECTION)
+        indexer.index_documents(SAMPLE_DOCS, collection_name=_COLLECTION)
     return db_path
 
 
@@ -57,7 +41,7 @@ def populated_db(tmp_path_factory):
 def app(populated_db):
     retriever = Retriever(
         dense_model=FakeEmbeddings(size=2048),
-        sparse_model=_FakeSparseEmbedding(),
+        sparse_model=FakeSparseEmbedding(),
         client=MilvusClient(populated_db),
         hybrid_search_weight=0.5,
         k=2,
