+++ tests/test_retriever.py
@@ -6,69 +6,33 @@ Nothing is mocked except SparseModel.from_config (avoids ~500 MB SPLADE download
 The reranker is disabled (reranker_model=None) to avoid the bge-reranker download.
 """
 
-from typing import Dict, List
 from unittest.mock import MagicMock, patch
 
 import pytest
 from langchain_community.embeddings import FakeEmbeddings
 from langchain_core.documents import Document
-from langchain_milvus.utils.sparse import BaseSparseEmbedding
 from pymilvus import MilvusClient
 
+from conftest import SAMPLE_DOCS, FakeSparseEmbedding
 from mmore.index.indexer import Indexer
 from mmore.rag.model import DenseModelConfig, SparseModelConfig
 from mmore.rag.retriever import Retriever
-from mmore.type import MultimodalSample
-
-
-class _FakeSparseEmbedding(BaseSparseEmbedding):
-    """Deterministic sparse embedder — no model download, runs on CPU."""
-
-    def embed_query(self, query: str) -> Dict[int, float]:
-        return {0: 1.0, 1: float(len(query))}
-
-    def embed_documents(self, texts: List[str]) -> List[Dict[int, float]]:
-        return [{0: 1.0, i + 1: float(len(t))} for i, t in enumerate(texts)]
 
 
 _COLLECTION = "test_col"
-
-_DOCS = [
-    MultimodalSample(
-        id="doc-1",
-        document_id="doc-1",
-        text="Paris is the capital of France.",
-        modalities=[],
-        metadata={},
-    ),
-    MultimodalSample(
-        id="doc-2",
-        document_id="doc-2",
-        text="The Eiffel Tower stands 330 metres tall.",
-        modalities=[],
-        metadata={"author": "Alice"},
-    ),
-    MultimodalSample(
-        id="doc-3",
-        document_id="doc-3",
-        text="Milvus is an open-source vector database.",
-        modalities=[],
-        metadata={},
-    ),
-]
 
 
 @pytest.fixture(scope="module")
 def populated_db(tmp_path_factory):
     """
-    Indexes _DOCS into Milvus Lite once and shares the db path across all
+    Indexes SAMPLE_DOCS into Milvus Lite once and shares the db path across all
     tests in this module — avoids re-indexing for every test.
     """
     db_path = str(tmp_path_factory.mktemp("retriever_db") / "test.db")
 
     with patch(
         "mmore.index.indexer.SparseModel.from_config",
-        return_value=_FakeSparseEmbedding(),
+        return_value=FakeSparseEmbedding(),
     ):
         client = MilvusClient(db_path, enable_sparse=True)
         indexer = Indexer(
@@ -78,7 +42,7 @@ def populated_db(tmp_path_factory):
             ),
             client=client,
         )
-        indexer.index_documents(_DOCS, collection_name=_COLLECTION)
+        indexer.index_documents(SAMPLE_DOCS, collection_name=_COLLECTION)
 
     return db_path
 
@@ -92,7 +56,7 @@ def retriever(populated_db):
     client = MilvusClient(populated_db)
     return Retriever(
         dense_model=FakeEmbeddings(size=2048),
-        sparse_model=_FakeSparseEmbedding(),
+        sparse_model=FakeSparseEmbedding(),
         client=client,
         hybrid_search_weight=0.5,
         k=2,
@@ -142,7 +106,7 @@ def test_retrieve_min_score_filters_all(retriever):
 
 
 def test_retrieve_result_ids_are_known_docs(retriever):
-    known_ids = {d.id for d in _DOCS}
+    known_ids = {d.id for d in SAMPLE_DOCS}
     results = retriever.retrieve("Paris", collection_name=_COLLECTION, k=3)
     for r in results:
         assert r["id"] in known_ids
@@ -196,7 +160,7 @@ def test_get_documents_by_ids_unknown_id(retriever):
 def test_list_files_returns_all_documents(retriever):
     files = retriever.list_files(collection_name=_COLLECTION)
     returned_ids = {f["id"] for f in files}
-    expected_ids = {d.document_id for d in _DOCS}
+    expected_ids = {d.document_id for d in SAMPLE_DOCS}
     assert returned_ids == expected_ids
